🚚 Rename bus route color json file and add campus culor
