+++ src/Components/NearbyItem.js
@@ -3,7 +3,7 @@ import { Actions } from 'react-native-router-flux';
 import { connect } from 'react-redux';
 import { View, Text, TouchableOpacity, Image } from 'react-native';
 import { getBusStops } from '../actions';
-import { routeColor } from '../../route_color.json';
+import { routeColor } from '../../bus_color.json';
 
 var color = 'rgb(142, 142, 147)';
 var tag = '?';
